Parse HAL JSON using hal.JSON(hal HAL)
A better way of parsing HAL into JSON. By creating
a generic function for parsing to JSON we are able
to create a better API for the developer.

However, the use of interface{} as input does not
seem very type safe, and it should be able to be
solved using Go interfaces maybe?

Also, I have not added proper error handling.